Fix requests errors not being handled by dealing with them explicitly
The requests lib exceptions base class RequestException can only be used
with a "try/except" to catch all the specific requests lib exceptions.
So the session context manager exit function will refer to the specific
exception and not the base class. In our case, as RequestException was
erroneously thought as a "catch-all" for these with its place in the
"errors_without_traceback" list, all requests lib exception could
not fail without traceback (and so huge verbosity). This is fixed now.